feat: 포인트 거래

+++ domain/src/main/java/gcu/mp/domain/member/domin/Member.java
@@ -81,7 +81,7 @@ public class Member extends BaseEntity {
         payHistoryList.add(payHistory);
     }
 
-    public void addPoint(int totalAmount) {
+    public void addPoint(long totalAmount) {
         this.point += totalAmount;
     }
